+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-03T14:57:42.920Z",
+    "2026-06-03T19:59:14.631Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {