+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -182,7 +182,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KT6ZVW2BHF70QD5V4JKHWAVT"),
+  _id: z.string().default("01KT7H3M8HRB540MQJFDWS8ZM5"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,